+++ src/components/StakingVaults/PositionTable.tsx
@@ -1,18 +1,16 @@
-import { ArrowDownIcon, ArrowUpIcon, Search2Icon } from '@chakra-ui/icons'
-import { Button, Circle, Flex, IconButton, Tag, useColorModeValue } from '@chakra-ui/react'
+import { ArrowDownIcon, ArrowUpIcon } from '@chakra-ui/icons'
+import { Flex, IconButton, Tag } from '@chakra-ui/react'
 import type { AssetId } from '@shapeshiftoss/caip'
 import { fromAssetId } from '@shapeshiftoss/caip'
 import { matchSorter } from 'match-sorter'
 import { useCallback, useMemo } from 'react'
 import { useTranslate } from 'react-polyglot'
-import { Link } from 'react-router-dom'
 import type { Column, Row } from 'react-table'
 import { Amount } from 'components/Amount/Amount'
 import { AssetIcon } from 'components/AssetIcon'
 import { PositionDetails } from 'components/EarnDashboard/components/PositionDetails/PositionDetails'
-import { DefiIcon } from 'components/Icons/DeFi'
 import { ReactTable } from 'components/ReactTable/ReactTable'
-import { RawText, Text } from 'components/Text'
+import { RawText } from 'components/Text'
 import { isEthAddress } from 'lib/address/utils'
 import { bnOrZero } from 'lib/bignumber/bignumber'
 import type { AggregatedOpportunitiesByAssetIdReturn } from 'state/slices/opportunitiesSlice/types'
@@ -24,6 +22,9 @@ import {
   selectOpportunityApiPending,
 } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
+
+import { ResultsEmpty } from './ResultsEmpty'
+import { SearchEmpty } from './SearchEmpty'
 
 export type RowProps = Row<AggregatedOpportunitiesByAssetIdReturn>
 
@@ -44,61 +45,6 @@ const AssetCell = ({ assetId }: { assetId: AssetId }) => {
             {`on ${networkName}`}
           </RawText>
         ) : null}
-      </Flex>
-    </Flex>
-  )
-}
-
-const ResultsEmpty = () => {
-  const bgColor = useColorModeValue('gray.100', 'gray.750')
-  const translate = useTranslate()
-  return (
-    <Flex p={6} textAlign='center' alignItems='center' width='full' flexDir='column' gap={4}>
-      <Flex>
-        <Circle bg={bgColor} size='40px'>
-          <DefiIcon />
-        </Circle>
-      </Flex>
-      <Flex alignItems='center' textAlign='center' flexDir='column' gap={2}>
-        <Text
-          fontWeight='bold'
-          fontSize='lg'
-          letterSpacing='0.02em'
-          translation='defi.noActivePositions'
-        />
-        <Text
-          color='gray.500'
-          letterSpacing='0.012em'
-          translation='assets.assetCards.stakingBody'
-        />
-        <Button as={Link} to='/defi/earn' mt={4}>
-          {translate('defi.startEarning')}
-        </Button>
-      </Flex>
-    </Flex>
-  )
-}
-const SearchEmpty = ({ searchQuery }: { searchQuery?: string }) => {
-  const bgColor = useColorModeValue('gray.100', 'gray.750')
-  return (
-    <Flex p={6} textAlign='center' alignItems='center' width='full' flexDir='column' gap={4}>
-      <Flex>
-        <Circle bg={bgColor} size='40px'>
-          <Search2Icon />
-        </Circle>
-      </Flex>
-      <Flex alignItems='center' textAlign='center' flexDir='column' gap={2}>
-        <Text
-          fontWeight='bold'
-          fontSize='lg'
-          letterSpacing='0.02em'
-          translation='common.noResultsFound'
-        />
-        <Text
-          color='gray.500'
-          letterSpacing='0.012em'
-          translation={['common.noResultsBody', { searchQuery: `"${searchQuery}"` }]}
-        />
       </Flex>
     </Flex>
   )
