+++ src/components/StakingVaults/ProviderTable.tsx
@@ -1,14 +1,5 @@
-import { ArrowDownIcon, ArrowUpIcon, Search2Icon } from '@chakra-ui/icons'
-import {
-  Avatar,
-  Button,
-  Circle,
-  Flex,
-  IconButton,
-  Link,
-  Tag,
-  useColorModeValue,
-} from '@chakra-ui/react'
+import { ArrowDownIcon, ArrowUpIcon } from '@chakra-ui/icons'
+import { Avatar, Flex, IconButton, Tag } from '@chakra-ui/react'
 import type { DefiProvider } from 'features/defi/contexts/DefiManagerProvider/DefiCommon'
 import { DefiProviderMetadata } from 'features/defi/contexts/DefiManagerProvider/DefiCommon'
 import { matchSorter } from 'match-sorter'
@@ -17,9 +8,8 @@ import { useTranslate } from 'react-polyglot'
 import type { Column, Row } from 'react-table'
 import { Amount } from 'components/Amount/Amount'
 import { ProviderDetails } from 'components/EarnDashboard/components/ProviderDetails/ProviderDetails'
-import { DefiIcon } from 'components/Icons/DeFi'
 import { ReactTable } from 'components/ReactTable/ReactTable'
-import { RawText, Text } from 'components/Text'
+import { RawText } from 'components/Text'
 import { bnOrZero } from 'lib/bignumber/bignumber'
 import type { AggregatedOpportunitiesByProviderReturn } from 'state/slices/opportunitiesSlice/types'
 import {
@@ -27,6 +17,9 @@ import {
   selectOpportunityApiPending,
 } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
+
+import { ResultsEmpty } from './ResultsEmpty'
+import { SearchEmpty } from './SearchEmpty'
 
 export type RowProps = Row<AggregatedOpportunitiesByProviderReturn>
 
@@ -41,62 +34,6 @@ const ProviderCell: React.FC<ProviderCellProps> = ({ provider }) => {
       <Avatar size='sm' src={icon} />
       <Flex flexDir='column'>
         <RawText>{type}</RawText>
-      </Flex>
-    </Flex>
-  )
-}
-
-const ResultsEmpty = () => {
-  const bgColor = useColorModeValue('gray.100', 'gray.750')
-  const translate = useTranslate()
-  return (
-    <Flex p={6} textAlign='center' alignItems='center' width='full' flexDir='column' gap={4}>
-      <Flex>
-        <Circle bg={bgColor} size='40px'>
-          <DefiIcon />
-        </Circle>
-      </Flex>
-      <Flex alignItems='center' textAlign='center' flexDir='column' gap={2}>
-        <Text
-          fontWeight='bold'
-          fontSize='lg'
-          letterSpacing='0.02em'
-          translation='defi.noActivePositions'
-        />
-        <Text
-          color='gray.500'
-          letterSpacing='0.012em'
-          translation='assets.assetCards.stakingBody'
-        />
-        <Button as={Link} to='/defi/earn' mt={4}>
-          {translate('defi.startEarning')}
-        </Button>
-      </Flex>
-    </Flex>
-  )
-}
-
-const SearchEmpty = ({ searchQuery }: { searchQuery?: string }) => {
-  const bgColor = useColorModeValue('gray.100', 'gray.750')
-  return (
-    <Flex p={6} textAlign='center' alignItems='center' width='full' flexDir='column' gap={4}>
-      <Flex>
-        <Circle bg={bgColor} size='40px'>
-          <Search2Icon />
-        </Circle>
-      </Flex>
-      <Flex alignItems='center' textAlign='center' flexDir='column' gap={2}>
-        <Text
-          fontWeight='bold'
-          fontSize='lg'
-          letterSpacing='0.02em'
-          translation='common.noResultsFound'
-        />
-        <Text
-          color='gray.500'
-          letterSpacing='0.012em'
-          translation={['common.noResultsBody', { searchQuery: `"${searchQuery}"` }]}
-        />
       </Flex>
     </Flex>
   )
